\name{geppe-package}
\alias{geppe-package}
\docType{package}
\title{
Generalised Exponential Poisson and Poisson Exponential Distributions
}
\description{
The package offers sime functions (including MLE) for the exponential-Poisson (EP), the generalised EP (GEP) and the Poisson-exponential (PE) distributions.
}
\details{
\tabular{lll}{
Package: \tab geppe\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2024-06-23\cr
License: \tab GPL-2\cr
}
}
\author{
Michail Tsagris \email{mtsagris@uoc.gr} and Sofia Piperaki \email{sofiapip23@gmail.com}.
}

\section{Maintainers}{
Michail Tsagris \email{mtsagris@uoc.gr}.
}

\references{
Barreto-Souza W. and Cribari-Neto F. (2009). A generalization of the exponential-Poisson distribution.
Statistics and Probability Letters, 79(24): 2493--2500.

Louzada F., Ramos P. L. and Ferreira H. P. (2020). Exponential-Poisson distribution:
estimation and applications to rainfall and aircraft data with zero occurrence.
Communications in Statistics-Simulation and Computation, 49(4): 1024--1043.

Rodrigues G. C., Louzada F. and Ramos P. L. (2018). Poisson-exponential distribution: different methods of estimation. Journal of Applied Statistics, 45(1): 128--144.
}



